pyproject.toml
[tool.poetry]
name = "assesment"
version = "0.1.0"
description = "assesment using crewAI"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
crewai = { extras = ["tools"], version = "^0.41.1" }
flask = "^3.0.3"
flask-bcrypt = "^1.0.1"
python-dotenv = "^1.0.1"
pypdf = "^4.3.1"
pymupdf = "^1.24.9"
groq = "^0.9.0"
langchain-mistralai = "^0.1.11"

[tool.poetry.scripts]
assesment = "assesment.main:run"
train = "assesment.main:train"
replay = "assesment.main:replay"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
